Refactor Quote - split and store api info in api.js
